Split IMinitwitRepositories into user and message
